static external to login routes + conversations integrate first layer
