+++ client-admin/package.json
@@ -70,7 +70,7 @@
     "boolean": "^0.1.3",
     "color": "^0.7.3",
     "highlight.js": "^8.9.1",
-    "history": "^1.17.0",
+    "history": "4.10.1",
     "jquery": "^2.1.4",
     "lodash": "^3.10.1",
     "moment": "^2.10.6",
