--- client-admin/src/components/framework/elephant.js
@@ -1,78 +1,0 @@
-// Copyright (C) 2012-present, The Authors. This program is free software: you can redistribute it and/or  modify it under the terms of the GNU Affero General Public License, version 3, as published by the Free Software Foundation. This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public License for more details. You should have received a copy of the GNU Affero General Public License along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-import React from "react";
-import Radium from "radium";
-import _ from "lodash";
-import Flex from "./flex";
-
-@Radium
-class Elephant extends React.Component {
-  styles() {
-    return {
-
-    }
-  };
-  render() {
-    return (
-      <svg width="221px" height="220px" viewBox="0 0 221 220">
-          <g id="Page-1" stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
-              <g id="news-0.0.1" transform="translate(-130.000000, -1266.000000)">
-                  <g id="Group-3" transform="translate(130.000000, 1266.000000)">
-                    <path d="M110,220 C170.751322,220 220,170.751322 220,110 C220,49.2486775 170.751322,0 110,0 C49.2486775,0 0,49.2486775 0,110 C0,170.751322 49.2486775,220 110,220 Z" id="Oval-26" fill="#FFBA0D"></path>
-                    <path d="M158.090444,208.583346 L142.127336,179.369143 L123.730146,144.663256 L127.259648,124.863078 L123.730146,120.739298 L131.575547,107.838835 L131.575547,105.124125 L136.066629,90.6358439 L134.319436,88.7700297 L138.578004,78.9914957 L136.066629,77.7989451 L141.023736,69.5098807 L122.598413,49.7458779 L98.1287471,46.7271143 L167.223252,46.7271143 C167.223252,46.7271143 208.251763,68.03925 215.448281,124.863078 C222.644799,181.686906 184.773482,215.406908 184.773482,215.406908 L151.729714,215.406909 L158.090444,208.583346 Z" id="Rectangle-25" fill="#CC9200" transform="translate(157.243293, 131.067011) rotate(32.000000) translate(-157.243293, -131.067011) "></path>
-                    <polygon id="Path-22" stroke="#FFFFFF" strokeWidth="1.20000005" fill="#989898" points="105.826683 196.977509 108.030949 198.965113 113.578535 198.965113 116.497292 196.977508 111.172302 184.374756"></polygon>
-                    <path d="M118.412024,168.601125 L111.175742,184.37434 C111.175742,184.37434 109.773092,181.054244 112.017956,186.367876 C114.26282,191.681508 116.718741,195.584117 116.718741,195.584117 L118.412024,168.601125 Z" id="Path-23" stroke="#FFFFFF" strokeWidth="1.20000005" fill="#989898"></path>
-                    <path d="M103.883467,168.593297 L111.184631,184.382507 L105.614689,195.299306 L103.883467,168.593297 Z" id="Path-24" stroke="#FFFFFF" strokeWidth="1.20000005" fill="#989898"></path>
-                    <polygon id="Path-25" stroke="#FFFFFF" stroke-width="1.20000005" fill="#BBBBBB" points="110.697025 142.655303 104.026334 168.822517 111.174909 184.489332 118.353393 168.717774"></polygon>
-                    <polygon id="Path-26" stroke="#FFFFFF" stroke-width="1.20000005" fill="#989898" points="103.852914 167.558176 110.821377 142.764749 101.199016 135.433202"></polygon>
-                    <polygon id="Path-28" stroke="#FFFFFF" stroke-width="1.20000005" fill="#989898" points="110.622045 143.142836 118.36683 168.211898 120.881838 136.37842"></polygon>
-                    <polygon id="Path-29" stroke="#FFFFFF" stroke-width="1.20000005" fill="#BBBBBB" points="110.516785 141.788269 98.4083733 97.9537083 100.789208 134.517134"></polygon>
-                    <polygon id="Path-30" stroke="#FFFFFF" stroke-width="1.20000005" fill="#BBBBBB" points="121.151471 135.460637 121.765756 102.860268 111.038269 141.679321"></polygon>
-                    <polygon id="Path-31" stroke="#FFFFFF" stroke-width="1.20000005" fill="#BBBBBB" points="95.4861285 70.2611842 99.6978133 125.024232 86.6409474 110.082075"></polygon>
-                    <polygon id="Path-32" stroke="#FFFFFF" stroke-width="1.20000005" fill="#BBBBBB" points="125.752817 70.6058351 134.979086 109.940613 121.893499 125.050381 122.431909 100.061908 124.384073 95.3842591"></polygon>
-                    <polygon id="Path-33" stroke="#FFFFFF" stroke-width="1.20000005" fill="#BBBBBB" points="97.6903507 94.2409957 106.783732 87.9198248 110.399136 141.792953"></polygon>
-                    <polygon id="Path-34" stroke="#FFFFFF" stroke-width="1.20000005" fill="#BBBBBB" points="115.095647 87.8225166 125.051941 94.8445633 111.04257 141.324845"></polygon>
-                    <polygon id="Path-35" stroke="#FFFFFF" stroke-width="1.20000005" fill="#BBBBBB" points="107.169535 87.9048213 115.391001 88.0527125 111.013421 141.591907"></polygon>
-                    <path id="Path-36" stroke="#FFFFFF" stroke-width="1.20000005" fill="#BBBBBB"></path>
-                    <polygon id="Path-37" stroke="#FFFFFF" stroke-width="1.20000005" fill="#FFEBBA" points="89.264888 113.444884 80.4163513 149.987286 96.084036 122.267278"></polygon>
-                    <polygon id="Path-38" stroke="#FFFFFF" stroke-width="1.20000005" fill="#FFEBBA" points="89.2799289 113.268116 73.392086 152.98696 72.6245715 164.431304 77.6824582 158.192652"></polygon>
-                    <polygon id="Path-39" stroke="#FFFFFF" stroke-width="1.20000005" fill="#FFEBBA" points="72.6619656 165.832283 77.816857 170.401829 77.693938 158.291378"></polygon>
-                    <polygon id="Path-40" stroke="#FFFFFF" stroke-width="1.20000005" fill="#FFEBBA" points="124.034607 122.657837 141.671814 158.769348 142.510281 144.960569 132.685761 113.151955"></polygon>
-                    <polygon id="Path-41" stroke="#FFFFFF" stroke-width="1.20000005" fill="#FFEBBA" points="137.378659 151.946626 133.658058 168.21391 141.460562 160.508266"></polygon>
-                    <polygon id="Path-42" stroke="#FFFFFF" stroke-width="1.20000005" fill="#FFEBBA" points="133.367615 169.702515 126.589886 186.414393 127.655357 189.81056 140.442932 162.854004"></polygon>
-                    <polygon id="Path-43" stroke="#FFFFFF" stroke-width="1.20000005" fill="#BBBBBB" points="87.9335702 104.235527 73.8725522 97.1721725 87.4310391 78.0193317 87.0196533 88.4484949 90.2778944 92.5566047"></polygon>
-                    <path id="Path-44" stroke="#FFFFFF" stroke-width="1.20000005" fill="#BBBBBB"></path>
-                    <polygon id="Path-45" stroke="#FFFFFF" stroke-width="1.20000005" fill="#919191" points="87.0560821 66.7419157 87.4764863 88.6149001 90.6582463 91.759791 95.5005547 70.2219807 91.3870106 64.0231747"></polygon>
-                    <path d="M81.7608098,50.7430972 L86.8056592,57.8463945 L86.8003922,77.197433 C86.8003922,77.197433 73.7862152,96.8462894 73.5202708,96.8380412 C72.53944,93.2918952 81.7608098,50.7430972 81.7608098,50.7430972 Z" id="Path-46" stroke="#FFFFFF" stroke-width="1.20000005" fill="#969696"></path>
-                    <polygon id="Path-47" stroke="#FFFFFF" stroke-width="1.20000005" fill="#BBBBBB" points="81.0593383 51.6691354 60.2081561 85.5978541 73.5528744 94.5786053"></polygon>
-                    <polygon id="Path-48" stroke="#FFFFFF" stroke-width="1.20000005" fill="#BBBBBB" points="54.2487602 77.2548914 60.1923551 83.7845621 80.9008488 51.2961345"></polygon>
-                    <polygon id="Path-49" stroke="#FFFFFF" stroke-width="1.20000005" fill="#BBBBBB" points="80.5024704 50.8857856 62.2474513 35.7804455 49.74402 45.5177326 46.4186139 69.992147 55.0488688 76.116257"></polygon>
-                    <polygon id="Path-50" stroke="#FFFFFF" stroke-width="1.20000005" fill="#919191" points="62.9395559 35.3388575 71.5403932 30.3697776 92.1531229 34.5359919 81.3672196 50.093817"></polygon>
-                    <polygon id="Path-51" stroke="#FFFFFF" stroke-width="1.20000005" fill="#BBBBBB" points="92.5826246 35.201552 82.031822 50.1163216 99.6806549 48.456252"></polygon>
-                    <polygon id="Path-52" stroke="#FFFFFF" stroke-width="1.20000005" fill="#BBBBBB" points="81.8240141 50.8450859 87.575602 58.2356753 99.318188 48.7593452"></polygon>
-                    <path d="M99.5681321,49.4028795 C99.5681321,49.4028795 87.6009795,58.2035946 87.6009795,58.2035943 C87.6009795,58.203594 87.5483093,65.8972769 87.5483093,65.8972769 L91.6719086,63.5390875 L99.5681321,49.4028795 Z" id="Path-53" stroke="#FFFFFF" stroke-width="1.20000005" fill="#BBBBBB"></path>
-                    <polygon id="Path-54" stroke="#FFFFFF" stroke-width="1.20000005" fill="#BBBBBB" points="100.535349 48.8752197 91.4521302 64.1079259 95.5005547 70.2167137 110.287479 57.3455486"></polygon>
-                    <polygon id="Path-55" stroke="#FFFFFF" stroke-width="1.20000005" fill="#BBBBBB" points="95.7270367 70.7482041 97.5570874 94.1275567 106.272572 88.0087137 110.501991 57.9258786"></polygon>
-                    <polygon id="Path-56" stroke="#FFFFFF" stroke-width="1.20000005" fill="#BBBBBB" points="110.980332 57.9790277 114.838186 87.8315502 106.702074 87.4963761"></polygon>
-                    <polygon id="Path-57" stroke="#FFFFFF" stroke-width="1.20000005" fill="#BBBBBB" points="111.389245 57.4393973 125.642764 70.6615376 124.233117 94.0614795 115.650744 87.3891204"></polygon>
-                    <polygon id="Path-58" stroke="#FFFFFF" stroke-width="1.20000005" fill="#BBBBBB" points="111.14744 57.3489003 122.133969 47.4148181 130.546361 63.9566187 125.779227 70.3785549"></polygon>
-                    <polygon id="Path-59" stroke="#FFFFFF" stroke-width="1.20000005" fill="#BBBBBB" points="93.0729365 34.4378338 99.9631588 48.4361416 111.0828 56.6459923 122.117211 46.5529418 127.725152 34.1773556 120.504544 31.755962 102.297407 31.846459"></polygon>
-                    <polygon id="Path-60" stroke="#FFFFFF" stroke-width="1.20000005" fill="#BBBBBB" points="128.095022 34.8120062 140.069137 49.8783479 123.229031 48.2182784 122.515589 46.9992022"></polygon>
-                    <polygon id="Path-61" stroke="#FFFFFF" stroke-width="1.20000005" fill="#BBBBBB" points="123.59485 48.589364 140.610203 50.0942958 134.060901 58.5119545"></polygon>
-                    <polygon id="Path-62" stroke="#FFFFFF" stroke-width="1.20000005" fill="#BBBBBB" points="123.259676 49.1155874 131.08647 64.3961756 135.332647 66.1836113 134.486572 59.421234"></polygon>
-                    <polygon id="Path-63" stroke="#FFFFFF" stroke-width="1.20000005" fill="#919191" points="125.903721 71.1197685 131.043855 64.4421424 135.678356 66.687809 134.741783 87.911992 130.936599 92.2007843"></polygon>
-                    <polygon id="Path-64" stroke="#FFFFFF" stroke-width="1.20000005" fill="#BBBBBB" points="131.154463 92.1734916 135.141598 87.8147915 135.237362 79.0724928 148.374272 97.3270331 133.970404 104.510293"></polygon>
-                    <path d="M140.414845,50.463945 L134.401342,58.4506655 L136.061412,66.2774601 L135.252684,79.1591592 C135.252684,79.1591592 147.756034,96.2221848 149.107165,96.9860817 C150.647341,96.4118134 140.414845,50.463945 140.414845,50.463945 Z" id="Path-65" stroke="#FFFFFF" stroke-width="1.20000005" fill="#969696"></path>
-                    <polygon id="Path-66" stroke="#FFFFFF" stroke-width="1.20000005" fill="#919191" points="128.122094 34.349252 151.304173 28.9012361 160.440062 35.3231723 140.87116 49.7471512"></polygon>
-                    <polygon id="Path-67" stroke="#FFFFFF" stroke-width="1.20000005" fill="#BBBBBB" points="140.593445 50.3791938 149.744656 94.6303179 160.966285 85.4417588"></polygon>
-                    <path d="M140.94442,49.7030998 L161.364664,35.4270763 L170.846261,44.0472759 L175.751295,69.8374881 L166.823214,74.7698147 L140.94442,49.7030998 Z" id="Path-68" stroke="#FFFFFF" stroke-width="1.20000005" fill="#BBBBBB"></path>
-                    <polygon id="Path-69" stroke="#FFFFFF" stroke-width="1.20000005" fill="#BBBBBB" points="141.039705 50.3959526 160.581793 83.3833108 168.961625 76.6156665"></polygon>
-                  </g>
-              </g>
-          </g>
-      </svg>
-    );
-  }
-}
-
-export default Elephant;
