--- client-admin/src/components/framework/govFlat.js
@@ -1,59 +1,0 @@
-// Copyright (C) 2012-present, The Authors. This program is free software: you can redistribute it and/or  modify it under the terms of the GNU Affero General Public License, version 3, as published by the Free Software Foundation. This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public License for more details. You should have received a copy of the GNU Affero General Public License along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-import React from "react";
-import Radium from "radium";
-import _ from "lodash";
-import Flex from "./flex";
-
-
-
-@Radium
-class GovFlat extends React.Component {
-  styles() {
-    return {
-
-    }
-  };
-  render() {
-    return (
-        <svg width="100" height="100" viewBox="0 0 250 250">
-          <g id="Page-1" stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
-            <g id="gov">
-              <g id="Group-2">
-                <ellipse id="Oval-29" fill="#F1360A" cx="125" cy="125" rx="125" ry="125"></ellipse>
-                <path d="M125.543478,53.8043478 L125.543478,33.1521739" id="Line" stroke="#979797" strokeWidth="3" strokeLinecap="square"></path>
-                  <g id="Group" transform="translate(52.173913, 29.347826)">
-                    <ellipse id="Oval-30" fill="#D8D8D8" cx="73.5805627" cy="56.667088" rx="35.9143223" ry="36.0149141"></ellipse>
-                    <rect id="Rectangle-18" fill="#939393" x="14.8913043" y="113.763903" width="117.378517" height="32.5012639"></rect>
-                    <rect id="Rectangle-26" fill="#939393" x="31.0965473" y="58.8631193" width="86.7199488" height="52.7047523"></rect>
-                    <polygon id="Rectangle-15" fill="#FFFFFF" points="41.3815809 58.8540971 52.9200325 58.8540971 52.9200322 104.979778 41.3815809 104.979778"></polygon>
-                    <rect id="Rectangle-15" fill="#FFFFFF" x="0" y="156.806117" width="148.913043" height="14.9330131"></rect>
-                    <rect id="Rectangle-15" fill="#FFFFFF" x="25.4028133" y="45.247725" width="98.1074169" height="14.9330131"></rect>
-                    <polygon id="Rectangle-15" fill="#FFFFFF" points="8.95152063 103.222952 72.5572236 99.6543892 137.732685 103.222952 137.732685 118.155966 72.5572236 114.612745 8.95152063 118.155966"></polygon>
-                    <rect id="Rectangle-15" fill="#FFFFFF" x="10.511509" y="146.265167" width="127.014066" height="5.27047523"></rect>
-                    <rect id="Rectangle-15" fill="#FFFFFF" x="20.4490263" y="113.763903" width="8.45762335" height="32.5012639"></rect>
-                    <rect id="Rectangle-15" fill="#FFFFFF" x="118.254476" y="113.763903" width="8.45762335" height="32.5012639"></rect>
-                    <rect id="Rectangle-15" fill="#FFFFFF" x="93.7276215" y="113.763903" width="8.45762335" height="32.5012639"></rect>
-                    <rect id="Rectangle-15" fill="#FFFFFF" x="69.2007673" y="113.763903" width="8.45762335" height="32.5012639"></rect>
-                    <rect id="Rectangle-15" fill="#FFFFFF" x="44.673913" y="113.763903" width="8.45762335" height="32.5012639"></rect>
-                    <polygon id="Rectangle-15" fill="#B7B7B7" points="10.5085041 151.370998 137.561851 151.370998 148.913043 156.806117 0 156.806117"></polygon>
-                    <polygon id="Rectangle-15" fill="#FFFFFF" points="67.6603532 58.8540971 79.1988049 58.8540971 79.1988045 104.979778 67.6603532 104.979778"></polygon>
-                    <polygon id="Rectangle-15" fillOpacity="0.667629076" fill="#676767" points="8.93419872 118.152377 73.3895204 114.61599 132.269827 117.399939 132.269821 120.791203 73.3895223 117.577929 14.883706 120.228945"></polygon>
-                    <polygon id="Rectangle-15" fill="#FFFFFF" points="95.6910438 58.8540971 107.229495 58.8540971 107.229495 104.979778 95.6910438 104.979778"></polygon>
-                    <polygon id="Rectangle-27" fill="#D8D8D8" points="73.5805627 99.7093023 137.732685 103.260866 137.732685 118.163105 73.5805627 114.642315"></polygon>
-                    <polygon id="Rectangle-32" fillOpacity="0.67" fill="#676767" points="25.4028133 60.1807381 117.816496 60.1807381 117.816496 62.8159757 30.7851043 61.9859002"></polygon>
-                    <ellipse id="Oval-31" fill="#979797" cx="73.3695652" cy="1.63043478" rx="1.63043478" ry="1.63043478"></ellipse>
-                    <rect id="Rectangle-33" fill="#C5C4C4" x="81.5217391" y="3.26086957" width="8.69565217" height="7.60869565"></rect>
-                    <rect id="Rectangle-33" fill="#EEEEEE" x="71.7391304" y="2.17391304" width="3.26086957" height="7.60869565"></rect>
-                    <rect id="Rectangle-33" fill="#EDEDED" x="76.0869565" y="2.17391304" width="7.60869565" height="7.60869565"></rect>
-                    <path d="M75.5434783,2.7173913 L75.5434783,9.23913043" id="Line" stroke="#AEAEAE" strokeLinecap="square"></path>
-                  </g>
-                </g>
-              </g>
-            </g>
-        </svg>
-    );
-  }
-}
-
-export default GovFlat;
