--- client-admin/src/components/framework/meditator.js
@@ -1,36 +1,0 @@
-// Copyright (C) 2012-present, The Authors. This program is free software: you can redistribute it and/or  modify it under the terms of the GNU Affero General Public License, version 3, as published by the Free Software Foundation. This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public License for more details. You should have received a copy of the GNU Affero General Public License along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-import React from "react";
-import Radium from "radium";
-import _ from "lodash";
-import Flex from "./flex";
-
-@Radium
-class Meditator extends React.Component {
-  styles() {
-    return {
-
-    }
-  };
-  render() {
-    return (
-      <svg width="221px" height="220px" viewBox="0 0 221 220">
-        <g id="Page-1" stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
-          <g id="news-0.0.1" transform="translate(-130.000000, -963.000000)">
-            <g id="Group" transform="translate(130.000000, 963.000000)">
-              <circle id="Oval-26" fill="#FFBA0D" cx="110" cy="110" r="110"></circle>
-              <path d="M103.132536,174.435486 C99.9879388,149.702047 77.8027702,116.911442 73.8818541,96.7797929 C71.1783625,82.898918 59.6505452,56.1453249 59.6505452,56.1453249 L185.253618,56.1453249 C185.253618,56.1453249 196.82201,67.8729302 204.309847,89.5208176 C211.797684,111.168705 213.367219,121.70954 210.646151,142.277393 C210.141221,146.094027 209.614562,147.578228 208.852221,151.050445 C199.521877,193.547138 172.142336,211.383608 172.142336,211.383608 L89.8853398,211.383608 C89.8853398,211.383608 105.902549,196.222682 103.132536,174.435486 Z" id="Rectangle-22" fill="#CC9200" transform="translate(135.771623, 133.764467) rotate(32.000000) translate(-135.771623, -133.764467) "></path>
-              <path d="M64.7876473,131.301916 C62.4659615,133.283426 61.5701782,135.102579 59.347394,136.414892 C53.5916023,139.813062 49.7254899,140.809397 49.7254899,140.809397 C49.7254899,140.809397 49.6698478,142.566204 49.9262253,143.946041 C50.1826028,145.325878 50.8131348,146.663159 50.8131348,146.663159 C50.8131348,146.663159 56.1249289,146.914769 61.9150497,145.485129 C65.3480776,144.637479 69.1577335,142.418652 71.5595468,141.333455 C75.9802571,139.336072 83.4333502,131.924558 83.4333502,131.924558 C83.4333502,131.924558 85.0697508,124.278238 85.7566904,119.995509 C86.4563628,115.633398 87.5931861,105.990036 87.5931861,105.990036 C87.5931861,105.990036 90.2978903,115.480942 90.5525098,119.995509 C90.8220972,124.775468 89.165807,133.873615 89.165807,133.873615 C89.165807,133.873615 88.0339672,137.833197 87.5931848,139.848393 C87.150763,141.871084 86.5161944,145.987277 86.5161944,145.987277 L81.762821,152.252247 C81.762821,152.252247 74.9919932,149.740146 71.3498325,148.643851 C68.0220698,147.642189 57.7116888,146.509278 57.7116888,146.509278 C57.7116888,146.509278 54.3098687,146.482729 51.7016783,147.138282 C49.1690894,147.774833 45.430713,149.834686 45.430713,149.834686 C45.430713,149.834686 42.7573672,151.502139 42.0716697,153.872155 C41.845873,154.65259 40.8989921,156.831071 40.7944715,158.335415 C40.5893349,161.287904 41.3667563,163.881179 41.3667563,163.881179 C41.3667563,163.881179 43.8206493,167.345734 47.4499864,170.399888 C50.1830967,172.69985 53.9192014,174.610386 55.8944313,175.509706 C57.5849381,176.279391 60.2404227,177.30974 62.3198959,178.121197 C66.3092542,179.677934 71.5668005,181.243687 78.2140601,182.108807 C93.3212194,184.074958 112.934703,183.584196 112.934703,183.584196 C112.934703,183.584196 141.029194,183.501684 152.768077,179.894856 C157.677297,178.386475 164.386203,176.535607 169.902968,172.615738 C173.538137,170.032813 176.414631,168.325628 178.246827,165.264818 C180.232845,161.947039 179.734058,157.413393 179.734058,157.413393 C179.734058,157.413393 178.359553,152.361845 175.763167,150.424891 C174.557215,149.525228 171.256163,147.9395 168.783497,147.138281 C166.492682,146.395988 159.432316,146.158885 159.432316,146.158885 C159.432316,146.158885 152.674438,147.42874 149.779789,148.302624 C145.863461,149.48495 139.112246,152.127008 139.112246,152.127008 C139.112246,152.127008 137.289515,150.414595 136.54581,149.525986 C135.684597,148.496974 134.29749,146.374147 134.29749,146.374147 L130.498849,129.190606 C130.498849,129.190606 129.879655,119.826897 130.302874,116.137374 C130.766579,112.094895 133.159622,105.9946 133.159622,105.9946 C133.159622,105.9946 133.819501,113.012011 134.29749,116.137374 C135.147604,121.695894 137.502489,131.924558 137.502489,131.924558 C137.502489,131.924558 143.846534,139.039668 148.879325,141.333455 C150.92067,142.263836 154.473245,144.284117 157.73082,145.139105 C163.755368,146.720316 169.692113,146.880679 169.692113,146.880679 C169.692113,146.880679 170.359956,145.639174 170.548748,144.719957 C170.780843,143.589892 170.954775,140.732835 170.954775,140.732835 C170.954775,140.732835 168.963484,140.290403 166.532257,139.171262 C164.674492,138.316096 162.577925,136.920124 160.809102,135.891436 C159.242589,134.980406 155.080099,130.727094 153.444177,129.396176 C150.474057,126.979809 147.649548,124.928438 147.649548,124.928438 C147.649548,124.928438 144.460872,119.545544 144.196685,106.296959 C143.971683,95.0134901 142.880646,84.951041 140.328933,81.3877857 C137.143935,77.7909163 135.023749,77.1344538 135.023749,77.1344538 C135.023749,77.1344538 132.878364,83.1903403 128.675519,88.208024 C126.398666,90.9263087 122.959637,92.916523 120.721084,94.8359753 C118.773127,96.5062558 110.471269,98.8423117 110.471269,98.8423117 C110.471269,98.8423117 106.706563,98.1447088 102.545484,96.1753236 C99.3907945,94.6822496 95.9962296,92.1559684 93.910954,90.3871262 C90.377519,87.3898778 85.3636155,77.5446584 85.3636155,77.5446584 C85.3636155,77.5446584 80.1095658,78.8335742 78.5287357,84.9922952 C78.1968464,86.285295 77.5105649,88.1576115 77.2124686,91.4089303 C76.5920507,98.1757895 76.3974352,108.3086 75.7737607,113.340386 C74.4665444,123.886964 73.2097926,124.894467 73.2097926,124.894467 C73.2097926,124.894467 68.4778539,128.152402 64.7876473,131.301916 Z" id="Path-19" fill="#40B9EE"></path>
-              <path d="M191.417821,140.791825 C190.856017,144.808329 192.587161,146.282867 192.58716,146.686538 C191.080403,147.718237 190.35172,147.494084 187.382907,148.207625 C185.053613,148.767459 182.175655,149.211392 180.224552,149.551743 C177.583443,150.012459 171.350159,150.391125 171.350159,150.391125 C171.350159,150.391125 170.323509,149.954497 169.868744,149.551743 C169.693456,149.396502 169.46,148.717142 169.46,148.717142 C169.46,148.717142 169.614158,148.272955 169.980685,147.981257 C170.200072,147.806659 171.217743,147.318256 171.217743,147.318256 L175.516398,146.641583 C175.516398,146.641583 175.136382,144.62876 175.412009,143.586075 C175.687637,142.543389 176.97759,141.113974 176.97759,141.113974 L177.6489,141.575814 L178.491815,141.910833 C178.491815,141.910833 178.303694,142.616249 178.118818,143.018519 C178.022723,143.227613 177.648901,143.744927 177.648901,143.744927 C177.648901,143.744927 177.81119,144.808329 177.81119,144.808329 C177.81119,144.808329 178.327433,145.208952 179.029283,145.253432 C180.244939,145.330475 181.942856,145.070174 181.942856,145.070174 L183.093399,142.37875 C183.093399,142.37875 180.093665,142.09597 178.399762,141.779827 C177.58997,141.62869 177.865235,141.639477 177.54645,141.493928 C176.560837,141.043925 177.196353,139.874778 177.417886,139.787468 C177.503279,139.753814 178.283574,139.622847 179.735251,139.54852 C182.049643,139.43002 185.412107,139.362745 185.412107,139.362745 C185.412107,139.362745 187.950081,140.622992 191.417821,140.791825 Z" id="Path-20" fill="#FFECBB" transform="translate(181.023580, 144.876935) scale(-1, 1) translate(-181.023580, -144.876935) "></path>
-              <path d="M49.9578208,140.791825 C49.3960172,144.808329 51.1271609,146.282867 51.1271603,146.686538 C49.620403,147.718237 48.8917205,147.494084 45.9229068,148.207625 C43.5936134,148.767459 40.7156546,149.211392 38.7645522,149.551743 C36.1234427,150.012459 29.8901595,150.391125 29.8901595,150.391125 C29.8901595,150.391125 28.8635091,149.954497 28.4087444,149.551743 C28.233456,149.396502 28,148.717142 28,148.717142 C28,148.717142 28.1541579,148.272955 28.5206854,147.981257 C28.7400724,147.806659 29.7577434,147.318256 29.7577434,147.318256 L34.0563982,146.641583 C34.0563982,146.641583 33.6763815,144.62876 33.9520094,143.586075 C34.2276373,142.543389 35.51759,141.113974 35.51759,141.113974 L36.1889005,141.575814 L37.0318154,141.910833 C37.0318154,141.910833 36.843694,142.616249 36.6588183,143.018519 C36.5627225,143.227613 36.1889008,143.744927 36.1889008,143.744927 C36.1889008,143.744927 36.3511898,144.808329 36.3511902,144.808329 C36.3511903,144.808329 36.8674329,145.208952 37.5692833,145.253432 C38.7849393,145.330475 40.4828564,145.070174 40.4828564,145.070174 L41.6333992,142.37875 C41.6333992,142.37875 38.6336648,142.09597 36.9397617,141.779827 C36.12997,141.62869 36.4052355,141.639477 36.0864502,141.493928 C35.1008374,141.043925 35.7363531,139.874778 35.9578864,139.787468 C36.043279,139.753814 36.8235745,139.622847 38.2752506,139.54852 C40.5896426,139.43002 43.9521067,139.362745 43.9521067,139.362745 C43.9521067,139.362745 46.4900809,140.622992 49.9578208,140.791825 Z" id="Path-20" fill="#FFECBB"></path>
-              <path d="M86.4805639,76.821559 C85.8590361,76.9785897 85.2,77.5605001 85.2,77.5605001 C85.2,77.5605001 88.4182693,85.7120736 92.6391453,89.40231 C93.744597,90.3687867 95.0434565,91.4005786 96.7728171,92.7498941 C97.8703765,93.6062534 99.0953797,94.5530006 100.570132,95.4254436 C105.092109,98.1005834 110.108974,98.9721434 110.108974,98.9721434 C110.108974,98.9721434 117.441544,97.2878734 120.798616,94.9849765 C123.046737,93.4428015 127.467794,90.20682 130.000562,86.9968594 C133.637658,82.387301 135.017949,77.1905136 135.017949,77.1905136 C135.017949,77.1905136 133.109142,76.115385 129.519003,76.3483919 C127.948122,75.0791711 125.95416,74.4218763 124.089351,73.2030943 C121.194631,71.3111951 118.728649,69.8735687 118.728649,69.8735687 C118.728649,69.8735687 118.004488,67.2100215 117.850077,65.8581501 C117.711631,64.6460625 117.850078,62.1816916 117.850078,62.1816916 C117.850078,62.1816916 118.843643,60.3958453 119.243825,59.4318042 C119.677563,58.3869247 120.351838,56.15493 120.351838,56.15493 C120.351838,56.15493 120.836309,56.1436309 121.151275,55.8971898 C121.712924,55.4577341 122.787186,53.9002444 122.787186,53.9002444 C122.787186,53.9002444 123.33114,52.3825967 123.403136,51.5916138 C123.532524,50.170105 123.490796,47.7311953 123.403136,47.284585 C123.315477,46.8379747 122.520713,47.5379573 122.520713,47.5379573 L122.787185,44.9966875 C122.787185,44.9966875 115.640831,42.5345972 113.376768,41.2247398 C110.339483,39.4675408 106.883139,35.7955182 106.883139,35.7955182 C106.883139,35.7955182 103.774577,38.343202 102.276886,39.3016953 C100.766367,40.268398 97.8585109,41.5711062 97.8585109,41.5711062 C97.8585109,41.5711062 97.6646317,43.8278769 97.6597256,44.9966872 C97.6562831,45.8168273 97.6597256,47.5379573 97.6597256,47.5379573 C97.6597256,47.5379573 97.1014431,46.0949874 96.7728168,47.9161084 C96.7478725,48.0543404 96.5731733,49.9892553 96.7728168,50.9990733 C97.1945364,53.1321748 97.6597254,54.2344468 97.6597254,54.2344468 L99.9612007,56.4237067 C99.9612007,56.4237067 100.711594,58.8201464 101.185795,59.8576525 C101.635763,60.8421381 102.733707,62.4896817 102.733707,62.4896817 C102.733707,62.4896817 102.639813,64.9745726 102.472592,66.1878143 C102.309528,67.370895 101.742852,69.678649 101.742852,69.678649 C101.742852,69.678649 97.4379552,72.7727321 95.3834309,73.9394217 C93.5228462,74.99598 89.9975251,76.3483927 89.9975251,76.3483927 C89.9975251,76.3483927 87.4582112,76.5745539 86.4805639,76.821559 Z" id="Path-22" fill="#FFECBB"></path>
-              <path d="M100.124883,34.9329774 C97.5463136,37.8779674 97.7434684,41.8045011 97.7434684,41.8045011 C97.7434684,41.8045011 98.7931567,41.6982889 101.617366,39.893077 C104.441575,38.0878651 106.883077,35.9625363 106.883077,35.9625363 C106.883077,35.9625363 110.79743,39.9187539 113.165942,41.278182 C116.087205,42.9548664 122.752402,45.0708739 122.752402,45.0708739 C122.752402,45.0708739 122.509826,38.3328758 120.657755,35.5659132 C118.805684,32.7989507 113.628956,31.4405989 113.628956,31.4405989 C113.628956,31.4405989 114.844983,30.3859623 114.005462,28.810392 C113.165942,27.2348218 110.497144,27.2348218 110.497144,27.2348218 C110.497144,27.2348218 107.614387,27.0046633 106.459622,28.3514597 C105.304857,29.6982561 106.735922,31.4044443 106.735922,31.4044443 C106.735922,31.4044443 102.703453,31.9879873 100.124883,34.9329774 Z" id="Path-24" fill="#40B9EE"></path>
-            </g>
-          </g>
-        </g>
-      </svg>
-    );
-  }
-}
-
-export default Meditator;
