--- client-admin/src/components/framework/upAndToTheRight.js
@@ -1,83 +1,0 @@
-// Copyright (C) 2012-present, The Authors. This program is free software: you can redistribute it and/or  modify it under the terms of the GNU Affero General Public License, version 3, as published by the Free Software Foundation. This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public License for more details. You should have received a copy of the GNU Affero General Public License along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-import React from "react";
-
-import _ from "lodash";
-import Flex from "./flex";
-
-class UpAndToTheRight extends React.Component {
-  styles() {
-    return {};
-  }
-  render() {
-    return (
-      <svg width="221px" height="220px" viewBox="0 0 221 220">
-        <g id="Page-1" stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
-          <g id="news-0.0.1" transform="translate(-130.000000, -1566.000000)">
-            <g id="Group-4" transform="translate(130.000000, 1566.000000)">
-              <circle id="Oval-26" fill="#FFBA0D" cx="110" cy="110" r="110"></circle>
-              <path
-                d="M56.08616,185.268351 C56.08616,185.268351 89.1547265,165.312954 92.9074522,161.564097 C109.00641,145.481738 116.092438,143.813533 124.636612,122.291591 C127.16086,115.933258 130.95044,104.813676 130.851571,93.9506558 C130.71119,78.5266456 125.794517,62.42384 124.636612,50.9119433 C123.365195,38.2714921 124.63661,37.4538983 124.63661,37.4538983 L174.16723,37.4538983 C174.16723,37.4538983 221.503045,70.7527035 219.278329,126.371486 C217.053613,181.990268 188.750253,202.093272 188.750253,202.093272 L59.2750631,202.093272 L56.08616,185.268351 Z"
-                id="Rectangle-23"
-                fill="#CC9200"
-                transform="translate(137.720261, 119.773585) rotate(32.000000) translate(-137.720261, -119.773585) "
-              ></path>
-              <path
-                d="M170.493303,146.5 L29.0430615,146.5"
-                id="Line"
-                stroke="#FFFFFF"
-                strokeWidth="3"
-                strokeLinecap="square"
-              ></path>
-              <path id="Path-21" stroke="#979797"></path>
-              <path
-                d="M32.2405014,132.341784 C32.2405014,132.341784 93.9839599,131.650929 109.186661,124.694829 C117.821468,120.743914 138.361886,106.149167 143.985955,95.8492248 C149.610024,85.5492828 165.686814,45.3119301 165.686814,45.3119301"
-                id="Path-22"
-                stroke="#FFFFFF"
-                strokeWidth="3"
-              ></path>
-              <circle
-                id="Oval-28"
-                stroke="#FFFFFF"
-                strokeWidth="2"
-                fill="#CC9200"
-                cx="33.5"
-                cy="131.5"
-                r="4.5"
-              ></circle>
-              <circle
-                id="Oval-28"
-                stroke="#FFFFFF"
-                strokeWidth="2"
-                fill="#CC9200"
-                cx="107.5"
-                cy="125.5"
-                r="4.5"
-              ></circle>
-              <circle
-                id="Oval-28"
-                stroke="#FFFFFF"
-                strokeWidth="2"
-                fill="#CC9200"
-                cx="143.5"
-                cy="97.5"
-                r="4.5"
-              ></circle>
-              <circle
-                id="Oval-28"
-                stroke="#FFFFFF"
-                strokeWidth="2"
-                fill="#CC9200"
-                cx="165.5"
-                cy="45.5"
-                r="4.5"
-              ></circle>
-            </g>
-          </g>
-        </g>
-      </svg>
-    );
-  }
-}
-
-export default UpAndToTheRight;
