--- client-admin/src/components/overall-stats.js
@@ -1,23 +1,0 @@
-// Copyright (C) 2012-present, The Authors. This program is free software: you can redistribute it and/or  modify it under the terms of the GNU Affero General Public License, version 3, as published by the Free Software Foundation. This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public License for more details. You should have received a copy of the GNU Affero General Public License along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-import React from "react";
-import { connect } from "react-redux";
-import Radium from "radium";
-import _ from "lodash";
-
-// @connect(state => state.data)
-@Radium
-class OverallStats extends React.Component {
-  render() {
-    return (
-      <div>
-        <h1>OverallStats</h1>
-        <div>
-          "OverallStats"
-        </div>
-      </div>
-    );
-  }
-}
-
-export default OverallStats;
