+++ client-admin/src/components/password-reset.js
@@ -4,24 +4,8 @@ import React from "react";
 import { connect } from "react-redux";
 import { doPasswordReset } from "../actions";
 import Radium from "radium";
-import Button from "./framework/generic-button";
-import LanderContainer from "./App/Container/LanderContainer";
-import ContainerInner from "./App/Container/ContainerInner";
-
-const styles = {
-  input: {
-    display: "block",
-    margin: "20px 0px",
-    color: "rgb(100,100,100)",
-    fontSize: 14,
-    padding: 7,
-    borderRadius: 3,
-    border: "1px solid rgb(200,200,200)",
-  },
-};
 
 @connect()
-@Radium
 class SignIn extends React.Component {
   handleClick(e) {
     e.preventDefault();
@@ -40,21 +24,14 @@ class SignIn extends React.Component {
 
   render() {
     return (
-      <LanderContainer>
-        <ContainerInner>
-          <h1>Password Reset</h1>
-          <form>
-            <input style={styles.input} ref="password" placeholder="new password" type="password" />
-            <input
-              style={styles.input}
-              ref="passwordRepeat"
-              placeholder="repeat new password"
-              type="password"
-            />
-            <Button onClick={this.handleClick.bind(this)}>Set new password</Button>
-          </form>
-        </ContainerInner>
-      </LanderContainer>
+      <div>
+        <h1>Password Reset</h1>
+        <form>
+          <input ref="password" placeholder="new password" type="password" />
+          <input ref="passwordRepeat" placeholder="repeat new password" type="password" />
+          <button onClick={this.handleClick.bind(this)}>Set new password</button>
+        </form>
+      </div>
     );
   }
 }
