+++ client-admin/src/store/index.js
@@ -2,7 +2,6 @@
 
 import { createStore, applyMiddleware, compose } from "redux";
 import thunk from "redux-thunk";
-import createHistory from "history/lib/createBrowserHistory";
 
 import rootReducer from "../reducers";
