+++ qa.md
@@ -1,0 +1,11 @@
+# QA Script for Admin Dash
+
+- User can see home page at `/home`
+- User is redirected to `/home` if not logged in
+- User can sign in at `/signin`
+- User can reset password at `/pwreset`
+- User can see social linkage at `/account`
+- User can see `/privacy` policy
+- User can see `/tos`
+- User can `/createuser`
+- User can get `/integrate` embed code for whole site
